+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/CombinationGraph/BasicCombinationGraphs/BarGraphWithLine.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
+import '@cerner/terra-graphs/lib/components/Bar/BarGraph.module.scss';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import getCombinationGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/combinationDefault';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/BarGraphWithLineData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(getCombinationGraphConfig('#barGraphWithLine'));
+const dataset = utils.deepClone(data);
+
+const BarGraphWithLineCombinationExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    graph.loadContent(Carbon.api.bar(dataset[0]));
+    graph.loadContent(Carbon.api.line(dataset[1]));
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="barGraphWithLine" />
+    </React.Fragment>
+  );
+};
+
+export default BarGraphWithLineCombinationExample;