+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Legend/LineWithLegendOption.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
+import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineDefault('#lineWithLegendOption'));
+
+const contentData = utils.deepClone(exampleData);
+contentData[0].showShapes = false;
+contentData[0].color = Carbon.helpers.COLORS.BLUE;
+contentData[0].style = { strokeDashArray: '2,2' };
+
+const LineWithLegendOptionGeneralExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    contentData.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="lineWithLegendOption" />
+    </React.Fragment>
+  );
+};
+
+export default LineWithLegendOptionGeneralExample;
+