+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Spline/ValueRegion.test.jsx
@@ -16,3 +16,4 @@ export default () => (
     <LineGraph graphID="simpleSpline" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
+